Rename navbar dependent components
