+++ src/app/app.module.ts
@@ -21,9 +21,9 @@ import { AnimesViewComponent } from "./views/animes-view/animes-view.component";
 import { MangasViewComponent } from "./views/mangas-view/mangas-view.component";
 import { ContactViewComponent } from "./views/contact-view/contact-view.component";
 import { SearchViewComponent } from "./views/search-view/search-view.component";
-import { NavbarItemLinkComponent } from "./components/ui/navbar-item-link/navbar-item-link.component";
+import { NavbarItemLinkComponent } from "./components/ui/navbar-item-link/navbar-link.component";
 import { MatInputModule } from "@angular/material/input";
-import { NavbarItemButtonComponent } from './components/ui/navbar-item-button/navbar-item-button.component';
+import { NavbarItemButtonComponent } from "./components/ui/navbar-item-button/navbar-button.component";
 
 @NgModule({
   declarations: [
