+++ src/app/components/ui/navbar-item-button/navbar-button.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input } from "@angular/core";
 
 @Component({
-  selector: "app-navbar-item-button",
-  templateUrl: "./navbar-item-button.component.html",
-  styleUrls: ["./navbar-item-button.component.sass"],
+  selector: "app-navbar-button",
+  templateUrl: "./navbar-button.component.html",
+  styleUrls: ["./navbar-button.component.sass"],
 })
 export class NavbarItemButtonComponent {
   @Input() callback = () => {};
